+++ packages/interactivity-router/CHANGELOG.md
@@ -8,6 +8,9 @@
 
 -   Preserve `media` attribute on intial style sheets after client-side navigation. ([70668](https://github.com/WordPress/gutenberg/pull/70668))
 
+-  Ignores `<noscript>` elements while preparing DOM. ([70905](https://github.com/WordPress/gutenberg/pull/70905))
+
+
 ## 2.26.0 (2025-06-25)
 
 ### New Features